# Parallelize fetch all processes tests (huge speedup)