Despite I recently implemented parallel tests on UNIX (#1709), `TestFetchAllProcesses` class is the slowest one to run because it gets all possible info for all processes in one go. In fact it's a singe unit-test, so it's not parallelized by the test runner. In here I used `multiprocessing.Pool` to do the trick.

On my main linux box (8 cores):

```
Before:
----------------------------------------------------------------------
Ran 1 test in 2.511s

After:
----------------------------------------------------------------------
Ran 1 test in 0.931s
```

On Windows (virtualized, 4 cores):

```
Before:
----------------------------------------------------------------------
Ran 1 test in 13.752s

After:
----------------------------------------------------------------------
Ran 1 test in 3.951s
```